fim: keep result order when adding new result
Previously, when using FIM providers with multiple
`minuet-n-completions`, the suggestion result would "flicker": the
user may see "aaaaaa [1/1]" at first, then after a fraction of second it changed to
"bbbbb [1/2]".

This is because the FIM completion function use multiple requests and
call minuet--display-suggestion multiple times with
increasing number of candidates, while inserting each new result at the
first position of the candidates.

In this PR, I modifed the "push" into "append" so that each result is
appended to the candidates, so that the first result user sees would
not change. Only "[1/1]" would change to "[1/2]"